+++ src/main/java/com/runningmate/runningmate/user/service/LoginService.java
@@ -1,0 +1,13 @@
+package com.runningmate.runningmate.user.service;
+
+import com.runningmate.runningmate.user.dto.UserLoginRequestDto;
+import com.runningmate.runningmate.user.entity.User;
+import java.util.Optional;
+
+
+public interface LoginService {
+
+    Optional<User> login(UserLoginRequestDto userLoginRequestDto);
+
+    void logout();
+}
