+++ src/main/java/com/runningmate/runningmate/user/service/UserService.java
@@ -1,21 +1,9 @@
 package com.runningmate.runningmate.user.service;
 
-import com.runningmate.runningmate.user.dto.UserLoginRequestDto;
 import com.runningmate.runningmate.user.dto.UserSaveDto;
-import com.runningmate.runningmate.user.entity.User;
-
-import java.util.Optional;
 
 
 public interface UserService {
 
-    Optional<User> login(UserLoginRequestDto userLoginRequestDto);
-
-    Optional<User> findByUserId(long loginUserId);
-
     void insertUser(UserSaveDto userSaveDto);
-
-    void loginUser(long userId);
-
-    void logout();
 }
